requirements.txt
mysql-connector
python3-memcached
